+++ src/IV.DX/IV.DX.Application.Contracts/Abstractions/DXUnitReadMapper.cs
@@ -2,9 +2,8 @@ using IV.DX.Kernel.Models;
 
 namespace IV.DX.Application.Contracts.Abstractions
 {
-    public abstract class DXUnitMapper<TRequest, TResponse, TUnit> where TUnit : DXUnit, new()
+    public abstract class DXUnitReadMapper<TResponse, TUnit> where TUnit : DXUnit, new()
     {
         public abstract Task<TResponse> ToDtoAsync(TUnit unit, CancellationToken ct = default);
-        public abstract Task<TUnit> ToUnitAsync(TRequest dto, CancellationToken ct = default);
     }
 }