+++ src/IV.DX/IV.DX.Application.Contracts/Abstractions/IDXUnitCommandService.cs
@@ -1,0 +1,8 @@
+namespace IV.DX.Application.Contracts.Abstractions
+{
+    public interface IDXUnitCommandService<TRequest>
+    {
+        Task<Guid> SaveAsync(TRequest dto, CancellationToken ct = default);
+        Task DeleteAsync(Guid id, CancellationToken ct = default);
+    }
+}