+++ src/IV.DX/IV.DX.Application.Contracts/Abstractions/IDXUnitDtoService.cs
@@ -1,11 +1,8 @@
 namespace IV.DX.Application.Contracts.Abstractions
 {
-    public interface IDXUnitDtoService<TDto>
+    public interface IDXUnitDtoService<TRequest, TResponse>
+        : IDXUnitQueryService<TResponse>
+        , IDXUnitCommandService<TRequest>
     {
-        Task<TDto?> GetAsync(Guid id, CancellationToken ct = default);
-        Task<IEnumerable<TDto>> GetAllAsync(CancellationToken ct = default);
-        Task<IEnumerable<TDto>> GetAsync(string filter, CancellationToken ct = default);
-        Task<Guid> SaveAsync(TDto dto, CancellationToken ct = default);
-        Task DeleteAsync(Guid id, CancellationToken ct = default);
     }
 }