+++ src/IV.DX/IV.DX.Application.Contracts/Abstractions/IDXUnitQueryService.cs
@@ -1,0 +1,9 @@
+namespace IV.DX.Application.Contracts.Abstractions
+{
+    public interface IDXUnitQueryService<TResponse>
+    {
+        Task<TResponse?> GetAsync(Guid id, CancellationToken ct = default);
+        Task<IEnumerable<TResponse>> GetAllAsync(CancellationToken ct = default);
+        Task<IEnumerable<TResponse>> GetAsync(string filter, CancellationToken ct = default);
+    }
+}